refactor: migrate redisClient to db
