+++ apps/api/src/helpers/middlewares/rateLimiter.ts
@@ -1,12 +1,12 @@
 import type { NextFunction, Request, Response } from 'express';
 
+import redisClient from '@hey/db/redisClient';
 import getIp from '@hey/helpers/getIp';
 import sha256 from '@hey/helpers/sha256';
 import rateLimit from 'express-rate-limit';
 import RedisStore from 'rate-limit-redis';
 
 import catchedError from '../catchedError';
-import redisClient from '../redisClient';
 
 const hashedIp = (req: Request): string => sha256(getIp(req)).slice(0, 25);
