+++ apps/api/src/routes/badges/hasHeyNft.ts
@@ -2,12 +2,12 @@ import type { Request, Response } from 'express';
 import type { Address } from 'viem';
 
 import { HEY_MEMBERSHIP_NFT_PUBLICATION_ID } from '@hey/data/constants';
+import { getRedis, setRedis } from '@hey/db/redisClient';
 import logger from '@hey/helpers/logger';
 import lensPg from 'src/db/lensPg';
 import catchedError from 'src/helpers/catchedError';
 import { CACHE_AGE_INDEFINITE } from 'src/helpers/constants';
 import { rateLimiter } from 'src/helpers/middlewares/rateLimiter';
-import { getRedis, setRedis } from 'src/helpers/redisClient';
 import { noBody } from 'src/helpers/responses';
 import { getAddress } from 'viem';
