+++ apps/api/src/routes/misc/verified.ts
@@ -1,10 +1,10 @@
 import type { Handler } from 'express';
 
+import { getRedis, setRedis } from '@hey/db/redisClient';
 import logger from '@hey/helpers/logger';
 import heyPg from 'src/db/heyPg';
 import catchedError from 'src/helpers/catchedError';
 import { CACHE_AGE_30_MINS, VERIFIED_FEATURE_ID } from 'src/helpers/constants';
-import { getRedis, setRedis } from 'src/helpers/redisClient';
 
 export const get: Handler = async (_, res) => {
   try {
