+++ apps/api/src/routes/oembed/index.ts
@@ -1,11 +1,11 @@
 import type { Handler } from 'express';
 
+import { getRedis, setRedis } from '@hey/db/redisClient';
 import logger from '@hey/helpers/logger';
 import sha256 from '@hey/helpers/sha256';
 import catchedError from 'src/helpers/catchedError';
 import { CACHE_AGE_1_DAY } from 'src/helpers/constants';
 import getMetadata from 'src/helpers/oembed/getMetadata';
-import { getRedis, setRedis } from 'src/helpers/redisClient';
 import { noBody } from 'src/helpers/responses';
 
 export const get: Handler = async (req, res) => {
