+++ apps/api/src/routes/preferences/get.ts
@@ -1,12 +1,12 @@
 import type { Preferences } from '@hey/types/hey';
 import type { Request, Response } from 'express';
 
+import { getRedis, setRedis } from '@hey/db/redisClient';
 import logger from '@hey/helpers/logger';
 import parseJwt from '@hey/helpers/parseJwt';
 import heyPg from 'src/db/heyPg';
 import catchedError from 'src/helpers/catchedError';
 import validateLensAccount from 'src/helpers/middlewares/validateLensAccount';
-import { getRedis, setRedis } from 'src/helpers/redisClient';
 import { noBody } from 'src/helpers/responses';
 
 export const get = [
