+++ apps/api/src/routes/preferences/update.ts
@@ -1,12 +1,12 @@
 import type { Request, Response } from 'express';
 
+import { delRedis } from '@hey/db/redisClient';
 import logger from '@hey/helpers/logger';
 import parseJwt from '@hey/helpers/parseJwt';
 import catchedError from 'src/helpers/catchedError';
 import { rateLimiter } from 'src/helpers/middlewares/rateLimiter';
 import validateLensAccount from 'src/helpers/middlewares/validateLensAccount';
 import prisma from 'src/helpers/prisma';
-import { delRedis } from 'src/helpers/redisClient';
 import { invalidBody, noBody } from 'src/helpers/responses';
 import { boolean, number, object, string } from 'zod';
