+++ apps/api/src/routes/tokens/all.ts
@@ -1,10 +1,10 @@
 import type { Handler } from 'express';
 
+import { getRedis, setRedis } from '@hey/db/redisClient';
 import logger from '@hey/helpers/logger';
 import heyPg from 'src/db/heyPg';
 import catchedError from 'src/helpers/catchedError';
 import { CACHE_AGE_1_DAY } from 'src/helpers/constants';
-import { getRedis, setRedis } from 'src/helpers/redisClient';
 
 export const get: Handler = async (_, res) => {
   try {
